Rename views -> view and qa -> searchSummary.
