+++ apps/searchSummary/src/view/SearchView.tsx
