+++ apps/searchSummary/src/view/SummaryUx.tsx
