+++ apps/searchSummary/src/view/chrome/AppHeader.tsx
