+++ apps/searchSummary/src/view/chrome/appHeader.scss
