+++ apps/searchSummary/src/view/controls/ExampleQuestion.tsx
