+++ apps/searchSummary/src/view/controls/ExampleQuestions.tsx
