+++ apps/searchSummary/src/view/controls/HistoryDrawer.tsx
