+++ apps/searchSummary/src/view/controls/SearchControls.tsx
