+++ apps/searchSummary/src/view/controls/exampleQuestions.scss
