+++ apps/searchSummary/src/view/controls/searchControls.scss
