+++ apps/searchSummary/src/view/progressReport/ProgressReport.tsx
