+++ apps/searchSummary/src/view/progressReport/SearchResultsDrawer.tsx
