+++ apps/searchSummary/src/view/progressReport/searchResultsDrawer.scss
