+++ apps/searchSummary/src/view/results/SearchResult.scss
