+++ apps/searchSummary/src/view/results/SearchResult.tsx
