+++ apps/searchSummary/src/view/results/SearchResultList.tsx
