+++ apps/searchSummary/src/view/searchView.scss
