+++ apps/searchSummary/src/view/summary/ConfidenceBadge.tsx
