+++ apps/searchSummary/src/view/summary/ConfidenceScore.tsx
