+++ apps/searchSummary/src/view/summary/SummaryCitation.tsx
