+++ apps/searchSummary/src/view/types.ts
