+++ sampleConfigurations/searchSummary-configuration.ts
